requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
CacheControl==0.12.11
cachetools==5.3.0
cairocffi==1.5.1
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chess==1.9.4
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
firebase-admin==6.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.9
msgpack==1.0.5
multidict==6.0.4
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
rsa==4.9
six==1.16.0
stockfish==3.28.0
tinycss2==1.2.1
uritemplate==4.1.1
urllib3==2.0.2
webencodings==0.5.1
yarl==1.9.2
